proj: update MSRV to 1.63, commit lockfile
This matches the other Rusticata crates and fixes a build error with
1.60 from `memchr` requiring 1.61+

Note: we have two transitive dependencies that have bumped their MSRV
within otherwise semver compatible versions:

* `winnow` 0.4.2+ requires an MSRV of 1.64.0[0]
* `toml_edit` 0.19.9+ requires an MSRV of 1.64.0[1]

Recently the Rust project guidance on commiting lockfiles for library
projects[2] was changed. They say:

"A lockfile is an appropriate way to pin versions for your project so
you can validate your MSRV"

As a result I've opted to commit a lockfile and use `--locked` in CI to
maintain MSRV. I think this is a better approach than manually working
around the issue with `cargo update --precise` and in line with current
best practices.

[0]: https://github.com/winnow-rs/winnow/blob/v0.6.6/CHANGELOG.md#042---2023-04-28
[1]: https://github.com/toml-rs/toml/blob/main/crates/toml_edit/CHANGELOG.md#0199---2023-05-18
[2]: https://blog.rust-lang.org/2023/08/29/committing-lockfiles.html